# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"shiharai kun service","description":"shiharai kun API service","version":"0.0.1"},"host":"localhost:8088","basePath":"/api","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/invoices":{"get":{"tags":["invoices"],"summary":"指定期間内に支払いが発生する請求書データの一覧を取得","operationId":"invoices#fetch list","parameters":[{"name":"user_id","in":"query","description":"ユーザID","required":true,"type":"string"},{"name":"from_date","in":"query","description":"指定する期間の開始日","required":true,"type":"string"},{"name":"to_date","in":"query","description":"指定する期間の最終日","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/Invoice"}}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/ErrBadRequest"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/ErrInternalServerError"}}},"schemes":["http"]},"post":{"tags":["invoices"],"summary":"請求書データを作成","operationId":"invoices#create","parameters":[{"name":"user_id","in":"query","description":"ユーザID","required":true,"type":"string"},{"name":"CreateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/InvoicesCreateRequestBody","required":["client_id","issue_date","payment_amount","payment_due_date"]}}],"responses":{"201":{"description":"Created response.","schema":{"$ref":"#/definitions/Invoice","required":["client_id","issue_date","payment_amount","commission","commission_rate","consumption_tax","consumption_tax_rate","billing_amount","payment_due_date","status"]}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/ErrBadRequest"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/ErrInternalServerError"}}},"schemes":["http"]}}},"definitions":{"ErrBadRequest":{"title":"Mediatype identifier: application/vnd.err.bad_request; view=default","type":"object","properties":{"message":{"type":"string","example":"不正なリクエスト"}},"description":"create_bad_request_response_body result type (default view)","example":{"message":"不正なリクエスト"},"required":["message"]},"ErrInternalServerError":{"title":"Mediatype identifier: application/vnd.err.internal_server_error; view=default","type":"object","properties":{"message":{"type":"string","example":"サーバーエラー"}},"description":"create_internal_server_error_response_body result type (default view)","example":{"message":"サーバーエラー"},"required":["message"]},"Invoice":{"title":"Invoice","type":"object","properties":{"billing_amount":{"type":"integer","description":"請求金額","example":10440,"format":"int64","minimum":1},"client_id":{"type":"string","description":"取引先ID","example":"Quod culpa ut accusantium nesciunt ut."},"commission":{"type":"integer","description":"手数料","example":400,"format":"int64","minimum":0},"commission_rate":{"type":"number","description":"手数料率","example":0.04,"format":"double","minimum":0},"consumption_tax":{"type":"integer","description":"消費税","example":40,"format":"int64","minimum":0},"consumption_tax_rate":{"type":"number","description":"消費税率","example":0.1,"format":"double","minimum":0},"issue_date":{"type":"string","description":"発行日","example":"2024-10-01"},"payment_amount":{"type":"integer","description":"支払金額","example":10000,"format":"int64","minimum":1},"payment_due_date":{"type":"string","description":"支払期日","example":"2024-10-27"},"status":{"type":"string","description":"ステータス","example":"未処理","enum":["未処理","処理中","支払い済み","エラー"]}},"example":{"billing_amount":10440,"client_id":"Et aliquam et a sit itaque.","commission":400,"commission_rate":0.04,"consumption_tax":40,"consumption_tax_rate":0.1,"issue_date":"2024-10-01","payment_amount":10000,"payment_due_date":"2024-10-27","status":"未処理"},"required":["client_id","issue_date","payment_amount","commission","commission_rate","consumption_tax","consumption_tax_rate","billing_amount","payment_due_date","status"]},"InvoicesCreateRequestBody":{"title":"InvoicesCreateRequestBody","type":"object","properties":{"client_id":{"type":"string","description":"取引先ID","example":"Neque exercitationem."},"issue_date":{"type":"string","description":"発行日","example":"2024-10-01"},"payment_amount":{"type":"integer","description":"支払金額","example":10000,"format":"int64","minimum":1},"payment_due_date":{"type":"string","description":"支払期日","example":"2024-10-27"}},"example":{"client_id":"Autem aperiam laboriosam et earum odio.","issue_date":"2024-10-01","payment_amount":10000,"payment_due_date":"2024-10-27"},"required":["client_id","issue_date","payment_amount","payment_due_date"]}}}